Update cp_macro.cfg
Updated the END_PRINT macro to disable ALL motors instead of just the FANS, Hotend and Bed. This is for safety reasons as users may forget their stepper motors are still active and helps avoid overheating motors.

Updated CANCEL_PRINT macro to include a filament cut and retraction of the active filament on job cancel. This avoids user starting a new print with filament loaded in the hotend. Afterwards, power is cut to ALL motors fan and hot items across all systems.